requirements.txt
llama-index==0.10.35
python-dotenv
transformers
tiktoken